requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-debug-toolbar==4.2.0
django-filter==23.3
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
email-validator==2.1.0.post1
Faker==0.7.4
Flask==2.3.3
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
Werkzeug==2.3.7
WTForms==3.1.0
